+++ test/test_helper.rb
@@ -12,6 +12,7 @@ require "resque"
 require "sidekiq"
 require "delayed_job_active_record"
 require "sidekiq/testing"
+require "torquebox-no-op"
 
 require "support/rails_app"
 require "support/helpers"